+++ src/renderer/src/hooks/use-window-stream-visibility.ts
@@ -1,35 +1,18 @@
 import { useEffect, useState, useSyncExternalStore } from 'react'
-import { isWindowVisible } from '@/lib/window-visibility-interval'
 import {
-  isDocumentVisibilityProvenStale,
-  registerStaleDocumentVisibilityRecovery
-} from '@/components/terminal-pane/stale-document-visibility'
-
-// Why the stale-latch term is load-bearing: macOS can wedge document.visibilityState at 'hidden'
-// with no further visibilitychange, so a window the user is looking at would park its stream
-// forever (same bug class as the field report of 78MB of terminal bytes dropped on visible ptys).
-function getWindowVisibleSnapshot(): boolean {
-  return isWindowVisible() || isDocumentVisibilityProvenStale()
-}
-
-function subscribeWindowVisible(onChange: () => void): () => void {
-  document.addEventListener('visibilitychange', onChange)
-  // Why: the stale latch flips to proven-visible without emitting a visibilitychange.
-  const unregisterStaleRecovery = registerStaleDocumentVisibilityRecovery(onChange)
-  return () => {
-    document.removeEventListener('visibilitychange', onChange)
-    unregisterStaleRecovery()
-  }
-}
+  getWindowParkVisible,
+  subscribeWindowParkVisibility,
+  WINDOW_HIDE_PARK_GRACE_MS
+} from '@/lib/window-park-visibility'
 
 // Avoid renegotiating expensive streams during a quick app-switch round trip.
-export const WINDOW_STREAM_PARK_DELAY_MS = 500
+export const WINDOW_STREAM_PARK_DELAY_MS = WINDOW_HIDE_PARK_GRACE_MS
 
 export function useWindowStreamVisible(parkDelayMs = WINDOW_STREAM_PARK_DELAY_MS): boolean {
   const rawVisible = useSyncExternalStore(
-    subscribeWindowVisible,
-    getWindowVisibleSnapshot,
-    getWindowVisibleSnapshot
+    subscribeWindowParkVisibility,
+    getWindowParkVisible,
+    getWindowParkVisible
   )
   const [effectiveVisible, setEffectiveVisible] = useState(rawVisible)
